refactor(Kafka Node): Wrap Kafka node in VersionedNodeType shell
